requirements.txt
discord.py==1.5.1
python-dotenv==0.14.0
tmdbv3api==1.6.2
youtube_dl==2020.9.20
youtube-search-python==1.3.1